renamed the shared folder into core, all files with shared folder relation was modified to work with core folder
