+++ mds-frontend/src/core/assets/MDSystem.png
