+++ mds-frontend/src/core/services/emr-service.js
