+++ mds-frontend/src/core/styles/index.css
