+++ mds-frontend/src/core/types/mdsystem__core.d.ts
