+++ mds-frontend/src/core/utils/data-transformer.js
