+++ mds-frontend/src/main.jsx
@@ -1,8 +1,8 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import '@shared/styles/index.css'
+import '@core/styles/index.css'
 import App from './App.jsx'
-import { RoleProvider } from '@shared/context/role-context.jsx'
+import { RoleProvider } from '@core/context/role-context.jsx'
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
