+++ mds-frontend/src/staff/components/layout/StaffSidebar.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link, useLocation } from 'react-router-dom';
-import logo from '@shared/assets/MDSystem.png';
+import logo from '@core/assets/MDSystem.png';
 
 /**
  * Staff Sidebar Navigation Component
